requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
pytz==2022.2.1
sqlparse==0.4.2
psycopg2-binary==2.8.6
google-cloud-bigquery==3.3.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
tzdata==2022.4
